refactor(#612): carve mcp.rs into src/mcp/{mod,log,tools,notifier}.rs -- pure move
Line-coverage-asserted mechanical split (the proven carve mechanism, 5th
application): the generator segments the 2,298-line source into 63
contiguous segments and asserts their concatenation reproduces the file
byte-for-byte, so every line is claimed exactly once. No logic changes.

- mcp/log.rs (122): per-PID log files, stderr redirection, mcp_trace,
  mcp_verbose, most_recent_mcp_log.
- mcp/tools.rs (700): tool_definitions, JSON-RPC response envelopes,
  truncate, handle_tool_call + the tool-call/truncate tests.
- mcp/notifier.rs (1,030): should_notify, replay_should_deliver, CDATA/
  attr escaping, poll constants, cwd repo resolution, cursor seeding,
  NotifierHeartbeat/NotifierHealth, run_notifier_loop + the notifier,
  replay, filter, and escaping tests.
- mcp/mod.rs (524): PROTOCOL_VERSION, MAX_LINE_BYTES, dispatch,
  run_stdio_loop, re-exports keeping every mcp:: path compiling, the
  shared cfg(test) testutil (make_tx, make_request, mcp_test_dir), and
  the protocol-level dispatch tests.

Coverage accounting: 2,275 of 2,298 lines copied exactly once (mod 449,
log 115, tools 678, notifier 1,006, testutil 27); 23 on the named
discard list (the old use header + its trailing blank, the tests-mod
wrapper and closing brace, one stale section-divider comment). Only
additions are the per-file module docs, use decls, mod decls, pub use
re-exports, and test-mod wrappers. Declared visibility fix-ups (private
items now crossing file boundaries) follow in the next commit; cargo fmt
deferred to its own commit so the moved lines stay verbatim-diffable.
